Update README.md
This repository is dedicated to showcasing a collection of examples utilizing GSAP (GreenSock Animation Platform). Here, you'll find a variety of demonstrations that illustrate the powerful capabilities of GSAP for creating high-performance animations and interactive web experiences. Whether you are a beginner looking to learn the basics or an experienced developer seeking inspiration for advanced techniques, this repository aims to provide valuable insights and practical examples to enhance your projects.